+++ src/SchemalessAttributes.php
@@ -6,8 +6,9 @@ use Countable;
 use ArrayAccess;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Contracts\Support\Arrayable;
 
-class SchemalessAttributes implements ArrayAccess, Countable
+class SchemalessAttributes implements ArrayAccess, Countable, Arrayable
 {
     /** @var \Illuminate\Database\Eloquent\Model */
     protected $model;
@@ -37,9 +38,9 @@ class SchemalessAttributes implements ArrayAccess, Countable
         return $this->get($name);
     }
 
-    public function get(string $name)
+    public function get(string $name, $default = null)
     {
-        return array_get($this->schemalessAttributes, $name);
+        return array_get($this->schemalessAttributes, $name, $default);
     }
 
     public function __set(string $name, $value)
@@ -94,6 +95,11 @@ class SchemalessAttributes implements ArrayAccess, Countable
     public function count()
     {
         return count($this->schemalessAttributes);
+    }
+
+    public function toArray(): array
+    {
+        return $this->all();
     }
 
     public static function scopeWithSchemalessAttributes(string $attributeName): Builder