+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `laravel-schemaless-attributes` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.1.2 - 2018-05-22
+
+- implement `Arrayable`
+
+## 1.1.1 - 2018-05-22
+
+- add return to schemalessAttributes macro
+
+
+## 1.1.0 - 2018-05-15
+
+- add support for setting default values
+
+## 1.0.0 - 2018-05-11
 
 - initial release